implementando uma parte do editar paciente, falta add o resto dos thfield e ver o problema do id na tela cadastro

+++ src/main/java/com/poseidon/dao/PacienteDao.java
@@ -8,7 +8,7 @@ import com.poseidon.model.Paciente;
 
 @Repository
 @Qualifier()
-public interface PacienteDao extends CrudRepository<Paciente, Long> {
+public interface PacienteDao extends CrudRepository<Paciente, Integer> {
 
 	public Paciente findByNome(String nome);
